bugfix: logic error in alignment scoring fixed.
The bug only affected scoring of reads from inside insertions. The intention was to adjust the scoring to account for any kind of reference skipping (soft-clipping, INS, DEL) if we were not 'strict'. However, when scoring reads from inside an insertion we never adjust for anything - the CIGAR is QI (Q being length of read) and position is at the insertion start point. The existing logic expression did not execute this intent.