enhancement: xmv code made faster
The xmv matrix can have 112 million elements when we look at longer indels. The bottle-neck in the old code was that I was using Pool.map over all contigs. The function processing each contig was creating an xmv matrix each time it was called and passing this pack across the process boundary (serializing and deserializing).

The functions have been rewritten to receive work via the queue, so we only create and transport as many xmv matrices as we have processes. The time taken to do this is still startling.